Encode the fixed array header and data block

+++ src/PureHDF/VOL/Native/FileFormat/Level2/ObjectHeaderMessages/DataLayout/DataLayoutMessage4.Writing.cs
@@ -17,7 +17,7 @@ internal partial record class DataLayoutMessage4
             (var indexingInformation, var encodeSize) = isFiltered
                 ? (
                     (IndexingInformation)new FixedArrayIndexingInformation(PageBits: 0), 
-                    FixedArrayHeader.ENCODE_SIZE
+                    (long)FixedArrayHeader.ENCODE_SIZE
                 )
                 : (
                     new ImplicitIndexingInformation(),
